# [opt](FileReader) merge small IO to optimize read performace