--- a/be/src/io/cache/block/cached_remote_file_reader.cpp
+++ b/be/src/io/cache/block/cached_remote_file_reader.cpp
@@ -59,7 +59,8 @@ Status CachedRemoteFileReader::close() {
 
 std::pair<size_t, size_t> CachedRemoteFileReader::_align_size(size_t offset,
                                                               size_t read_size) const {
-    size_t segment_size = std::min(std::max(read_size, (size_t)4096), // 4k
+    size_t min_size = 1024 * 1024; // 1MB;
+    size_t segment_size = std::min(std::max(read_size, min_size),
                                    (size_t)config::file_cache_max_file_segment_size);
     segment_size = BitUtil::next_power_of_two(segment_size);
     size_t left = offset;