--- a/be/src/io/cache/block/cached_remote_file_reader.h
+++ b/be/src/io/cache/block/cached_remote_file_reader.h
@@ -53,6 +53,8 @@ public:
 
     FileSystemSPtr fs() const override { return _remote_file_reader->fs(); }
 
+    FileReader* get_remote_reader() { return _remote_file_reader.get(); }
+
 protected:
     Status read_at_impl(size_t offset, Slice result, size_t* bytes_read,
                         const IOContext* io_ctx) override;