--- a/be/src/vec/exec/format/json/new_json_reader.cpp
+++ b/be/src/vec/exec/format/json/new_json_reader.cpp
@@ -41,9 +41,7 @@
 #include "common/compiler_util.h" // IWYU pragma: keep
 #include "exprs/json_functions.h"
 #include "io/file_factory.h"
-#include "io/fs/broker_file_reader.h"
 #include "io/fs/buffered_reader.h"
-#include "io/fs/s3_file_reader.h"
 #include "io/fs/stream_load_pipe.h"
 #include "runtime/descriptors.h"
 #include "runtime/runtime_state.h"
@@ -373,22 +371,14 @@ Status NewJsonReader::_open_file_reader() {
     _current_offset = start_offset;
     _file_description.start_offset = start_offset;
 
-    io::FileReaderSPtr json_file_reader;
     if (_params.file_type == TFileType::FILE_STREAM) {
-        RETURN_IF_ERROR(FileFactory::create_pipe_reader(_range.load_id, &json_file_reader));
+        RETURN_IF_ERROR(FileFactory::create_pipe_reader(_range.load_id, &_file_reader));
     } else {
         io::FileCachePolicy cache_policy = FileFactory::get_cache_policy(_state);
-        RETURN_IF_ERROR(FileFactory::create_file_reader(_profile, _system_properties,
-                                                        _file_description, &_file_system,
-                                                        &json_file_reader, cache_policy));
-    }
-    if (typeid_cast<io::S3FileReader*>(json_file_reader.get()) != nullptr ||
-        typeid_cast<io::BrokerFileReader*>(json_file_reader.get()) != nullptr) {
-        // PrefetchBufferedReader now only support csv&json format when reading s3&broker file
-        _file_reader.reset(
-                new io::PrefetchBufferedReader(json_file_reader, _range.start_offset, _range.size));
-    } else {
-        _file_reader = std::move(json_file_reader);
+        RETURN_IF_ERROR(io::DelegateReader::create_file_reader(
+                _profile, _system_properties, _file_description, &_file_system, &_file_reader,
+                io::DelegateReader::AccessMode::SEQUENTIAL, cache_policy, _io_ctx,
+                io::PrefetchRange(_range.start_offset, _range.size)));
     }
     return Status::OK();
 }