@morningman 2023-04-21
run buildall